+++ tests/lantern_tests/outputs/test_site_pages.py
@@ -1,4 +1,5 @@
 import logging
+from http import HTTPStatus
 from pathlib import Path
 
 from lantern.models.site import ExportMeta
@@ -13,9 +14,8 @@ class TestSitePagesOutput:
         """Can create a site pages output."""
         output = SitePagesOutput(logger=fx_logger, meta=fx_export_meta)
         assert isinstance(output, SitePagesOutput)
-        assert output.name == "Site Pages"
 
-    def test_outputs(self, fx_logger: logging.Logger, fx_export_meta: ExportMeta):
+    def test_content(self, fx_logger: logging.Logger, fx_export_meta: ExportMeta):
         """Can generate site content items."""
         expected_paths = [
             Path("404.html"),
@@ -28,11 +28,19 @@ class TestSitePagesOutput:
         ]
 
         output = SitePagesOutput(logger=fx_logger, meta=fx_export_meta)
-        outputs = _index_site_content_outputs(output.outputs)
+        content = _index_site_content_outputs(output.content)
 
-        assert len(outputs) > 1
+        assert len(content) > 1
         for path in expected_paths:
-            assert path in outputs
-            result = outputs[path]
+            assert path in content
+            result = content[path]
             assert result.object_meta == {"build_key": fx_export_meta.build_key}
             assert result.media_type == "text/html"
+
+    def test_checks(self, fx_logger: logging.Logger, fx_export_meta: ExportMeta):
+        """Can generate additional checks for 404 error handling."""
+        output = SitePagesOutput(logger=fx_logger, meta=fx_export_meta)
+        content = output.content
+        checks = output.checks
+        assert len(checks) == len(content) + 1
+        assert checks[-1].http_status == HTTPStatus.NOT_FOUND